k8s: Remove node annotation on startup
Node annotations have been fully replaced by the CiliumNode custom resource.
Remove the logic to annotate the node on startup. This avoids a patch operation
on the node when the agent starts up and simplifies understanding where values
such as health IPs, encryption keys and PodCIDRs are being derived from.

The remaining functionality in the controller which was annotating the node is
to patch the status of the node. Keep this but rename the controller and move
it to the end of the bootstrap phase to only mark the node as ready once the
bootstrapping has been complete. Change the k8s client tests around node
patching to use SetNetworkUnavailableFalse() instead.

The remaining functionality in annotate.go is related to secrets retrieval so
rename the file to secrets.go.

Signed-off-by: Thomas Graf <thomas@cilium.io>